+++ Bugsnag/KSCrash/Source/KSCrash/Recording/BSG_KSSystemInfo.m
@@ -24,19 +24,22 @@
 // THE SOFTWARE.
 //
 
-#import "BSGKeys.h"
 #import "BSG_KSSystemInfo.h"
-#import "BSG_KSSystemInfoC.h"
-#import "BSG_KSMachHeaders.h"
+
+#import "BSGKeys.h"
+#import "BSG_Jailbreak.h"
+#import "BSG_KSCrash.h"
+#import "BSG_KSCrashC.h"
+#import "BSG_KSCrashReportFields.h"
 #import "BSG_KSFileUtils.h"
 #import "BSG_KSJSONCodecObjC.h"
-#import "BSG_KSMach.h"
-#import "BSG_KSSysCtl.h"
-#import "BugsnagCollections.h"
 #import "BSG_KSLogger.h"
-#import "BSG_KSCrashReportFields.h"
 #import "BSG_KSMach.h"
-#import "BSG_KSCrash.h"
+#import "BSG_KSMach.h"
+#import "BSG_KSMachHeaders.h"
+#import "BSG_KSSysCtl.h"
+#import "BSG_KSSystemInfoC.h"
+#import "BugsnagCollections.h"
 
 #import <CommonCrypto/CommonDigest.h>
 #import <mach-o/dyld.h>
@@ -44,7 +47,6 @@
 #if TARGET_OS_IOS || TARGET_OS_TV
 #import "BSGUIKit.h"
 #endif
-#import "BSG_Jailbreak.h"
 
 
 static inline bool is_jailbroken() {
@@ -444,7 +446,12 @@ static NSDictionary * bsg_systemversion() {
         }
     }
 
-    NSDictionary *statsInfo = [[BSG_KSCrash sharedInstance] captureAppStats];
+    BSG_KSCrash_State state = crashContext()->state;
+    bsg_kscrashstate_updateDurationStats(&state);
+    NSMutableDictionary *statsInfo = [NSMutableDictionary dictionary];
+    statsInfo[@ BSG_KSCrashField_ActiveTimeSinceLaunch] = @(state.foregroundDurationSinceLaunch);
+    statsInfo[@ BSG_KSCrashField_BGTimeSinceLaunch] = @(state.backgroundDurationSinceLaunch);
+    statsInfo[@ BSG_KSCrashField_AppInFG] = @(state.applicationIsInForeground);
     sysInfo[@BSG_KSCrashField_AppStats] = statsInfo;
     return sysInfo;
 }
@@ -463,60 +470,6 @@ static NSDictionary * bsg_systemversion() {
     // Each Xcode app extension template is preconfigured with the appropriate extension point identifier key.
     return NSBundle.mainBundle.infoDictionary[@"NSExtension"][@"NSExtensionPointIdentifier"] != nil;
 }
-
-#if TARGET_OS_IOS || TARGET_OS_TV
-
-+ (UIApplicationState)currentAppState {
-    // Only checked outside of app extensions since sharedApplication is
-    // unavailable to extension UIKit APIs
-    if ([self isRunningInAppExtension]) {
-        return UIApplicationStateActive;
-    }
-
-    UIApplication * (^ getSharedApplication)(void) = ^() {
-        // Calling this API indirectly to avoid a compile-time check that
-        // [UIApplication sharedApplication] is not called from app extensions
-        // (which is handled above)
-        return [UIAPPLICATION performSelector:@selector(sharedApplication)];
-    };
-
-    __block UIApplication *application = nil;
-    if ([[NSThread currentThread] isMainThread]) {
-        application = getSharedApplication();
-    } else {
-        // [UIApplication sharedApplication] is a main thread-only API
-        dispatch_sync(dispatch_get_main_queue(), ^{
-            application = getSharedApplication();
-        });
-    }
-    
-    // There will be no UIApplication if UIApplicationMain() has not yet been
-    // called. This happens if started from a SwiftUI app's init() function or
-    // UIKit app's main() function. Returning UIApplicationStateActive (0) would
-    // be higly misleading, so we must check for this condition.
-    if (!application) {
-        return UIApplicationStateBackground;
-    }
-    
-    return application.applicationState;
-}
-
-+ (BOOL)isInForeground:(UIApplicationState)state {
-    // The app is in the foreground if the current state is "active" or
-    // "inactive". From the UIApplicationState docs:
-    // > UIApplicationStateActive
-    // >   The app is running in the foreground and currently receiving events.
-    // > UIApplicationStateInactive
-    // >   The app is running in the foreground but is not receiving events.
-    // >   This might happen as a result of an interruption or because the app
-    // >   is transitioning to or from the background.
-    // > UIApplicationStateBackground
-    // >   The app is running in the background.
-    return state == UIApplicationStateInactive
-        || state == UIApplicationStateActive;
-}
-
-#endif
 
 @end